v2/educational_app/requirements.txt
Flask==2.0.2
mysql-connector-python==8.0.26
python-dotenv==0.19.1
SQLAlchemy==1.4.25
Flask-SQLAlchemy==2.5.1
